% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as.ts.bcch_series}
\alias{as.ts.bcch_series}
\title{as.ts generic method for bcch_series objects}
\usage{
\method{as.ts}{bcch_series}(x, ...)
}
\arguments{
\item{x}{A \code{bcch_series} object}

\item{...}{Arguments passed to methods.}
}
\description{
as.ts generic method for bcch_series objects
}
\examples{

options(
 bcc_api_user = 178956728,
 bcc_api_pass = "cxynr4qyLLBw"
)

codigo <- "F032.PIB.FLU.R.CLP.EP18.Z.Z.0.T"

dfserie <- bcch_GetSeries(timeseries = codigo, firstdate = as.Date("2015-01-01"))

serie_ts <- as.ts(dfserie)

plot(serie_ts)

}
