% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{autoplot.bcch_series}
\alias{autoplot.bcch_series}
\title{autoplot generic method for bcch_series objects}
\usage{
\method{autoplot}{bcch_series}(object, ...)
}
\arguments{
\item{object}{A \code{bcch_series} object from \code{bcch_GetSeries}.}

\item{...}{Extra \code{bcch_series} objects.}
}
\description{
autoplot generic method for bcch_series objects
}
\examples{

options(
  bcc_api_user = 178956728,
  bcc_api_pass = "cxynr4qyLLBw"
)

fd <- lubridate::floor_date(Sys.Date(), unit = "year")

x <- bcch_GetSeries("F049.DES.TAS.INE9.10.M", firstdate = fd)

autoplot(x)

x2 <- bcch_GetSeries("F073.TCO.PRE.Z.D", firstdate = fd)

autoplot(x2)

autoplot(x, x2)

autoplot(x, x2) +
  ggplot2::facet_wrap(~descripEsp, scales = "free_y")

}
