% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{bcch_CatalogoSeries}
\alias{bcch_CatalogoSeries}
\title{Descarga y leer Catálogo de Series}
\usage{
bcch_CatalogoSeries()
}
\description{
Descarga y leer Catálogo de Series
}
\details{
https://si3.bcentral.cl/estadisticas/Principal1/web_services/index.htm
}
\examples{

bcch_CatalogoSeries()

}
