% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcchseries}
\alias{bcchseries}
\title{Series del}
\format{
A \code{data frame} with 19408 observations and 4 variables.
}
\usage{
bcchseries
}
\description{
This data comes from the \url{https://www.highcharts.com/} examples:
https://www.highcharts.com/demo/3d-area-multiple
}
\section{Variables}{


\itemize{

\item \code{seriesId}: seriesId.
\item \code{frequencyCode}: frequencyCode.
\item \code{spanishTitle}: spanishTitle.
\item \code{englishTitle}: englishTitle.

}
}

\keyword{datasets}
