% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{describe_series}
\alias{describe_series}
\title{Describir una serie del Banco Central}
\usage{
describe_series(
  series_id,
  token = NULL,
  verbose = getOption("bcchr.verbose", TRUE)
)
}
\arguments{
\item{series_id}{Vector de codigos exactos de series del Banco Central.}

\item{token}{Token de acceso a la API REST del Banco Central.}

\item{verbose}{Si es \code{TRUE}, informa cuando se consultan las cuatro
frecuencias. Por defecto usa la opcion \code{bcchr.verbose} y luego \code{TRUE}.}
}
\value{
Un tibble con una fila por cada codigo solicitado.
}
\description{
Devuelve la metadata de una o varias series identificadas por sus codigos
exactos. Esta funcion no interpreta nombres en lenguaje natural. Use
\code{\link[=resolve_series]{resolve_series()}} primero cuando no conozca los
codigos.
}
\examples{
if (nzchar(Sys.getenv("BCCH_TOKEN"))) {
  describe_series(
    c("F073.TCO.PRE.Z.D", "F073.UFF.PRE.Z.D"),
    verbose = FALSE
  )[c("series_id", "frequency", "spanish_title")]
}
}
