% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{bcch_GetSeries}
\alias{bcch_GetSeries}
\title{Bbtener los datos de series.}
\usage{
bcch_GetSeries(
  timeseries = NULL,
  user = getOption("bcc_api_user"),
  pass = getOption("bcc_api_pass"),
  firstdate = NULL,
  lastdate = NULL
)
}
\arguments{
\item{timeseries}{(obligatorio): incorporar código de la serie de tiempo a consultar.}

\item{user}{(obligatorio): completar nombre de usuario.}

\item{pass}{(obligatorio): completar contraseña.}

\item{firstdate}{(opcional): ingresar fecha desde la cual se requiere recoger datos. Si el parámetro no está presente, se recoge por defecto desde el primer dato disponible.}

\item{lastdate}{(opcional): ingresar fecha hasta la cual se requiere recoger datos. Si el parámetro no está presente, se recoge por defecto hasta el último dato disponible.}
}
\description{
Bbtener los datos de series.
}
\details{
https://si3.bcentral.cl/estadisticas/Principal1/web_services/index.htm
}
\examples{

options(
 bcc_api_user = 178956728,
 bcc_api_pass = "cxynr4qyLLBw"
)

bcch_GetSeries("F049.DES.TAS.INE9.10.M")

bcch_GetSeries("F073.TCO.PRE.Z.D")

}
