% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{bcch_GetSeries}
\alias{bcch_GetSeries}
\title{Obtener los datos de series.}
\usage{
bcch_GetSeries(
  timeseries = NULL,
  user = getOption("bcc_api_user"),
  pass = getOption("bcc_api_pass"),
  firstdate = NULL,
  lastdate = NULL
)
}
\arguments{
\item{timeseries}{Código de la serie de tiempo a consultar (obligatorio).}

\item{user}{Nombre de usuario (obligatorio).}

\item{pass}{Contraseña (obligatorio).}

\item{firstdate}{Fecha desde la cual se requiere recoger datos. Si el
parámetro no está presente, se recoge por defecto desde el primer
dato disponible (opcional).}

\item{lastdate}{Fecha hasta la cual se requiere recoger datos. Si el
parámetro no está presente, se recoge por defecto hasta el
último dato disponible (opcional).}
}
\description{
Obtener los datos de series.
}
\details{
https://si3.bcentral.cl/estadisticas/Principal1/web_services/index.htm
}
\examples{

options(
 bcc_api_user = 178956728,
 bcc_api_pass = "cxynr4qyLLBw"
)



}
