% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{bcch_SearchSeries}
\alias{bcch_SearchSeries}
\title{Obtener lista de series disponibles por frecuencia y su metadata.}
\usage{
bcch_SearchSeries(
  frequency = NULL,
  user = getOption("bcc_api_user"),
  pass = getOption("bcc_api_pass")
)
}
\arguments{
\item{frequency}{(obligatorio): completar con la frecuencia para la cual se quiere consultar el catálogo de series disponibles. Puede tomar los valores DAILY, MONTHLY, QUARTERLY o ANNUAL.}

\item{user}{(obligatorio): completar nombre de usuario.}

\item{pass}{(obligatorio): completar contraseña.}
}
\description{
Obtener lista de series disponibles por frecuencia y su metadata.
}
\details{
https://si3.bcentral.cl/estadisticas/Principal1/web_services/index.htm
}
\examples{

options(
 bcc_api_user = 178956728,
 bcc_api_pass = "cxynr4qyLLBw"
)

bcch_SearchSeries("ANNUAL")


}
